% Generated by roxygen2 (4.0.1): do not edit by hand
\name{effectsizeShape}
\alias{effectsizeShape}
\title{Effect size for multivariate shape vector}
\usage{
effectsizeShape(qtl, shape = NULL)
}
\arguments{
\item{qtl}{A qtl effect obtained from fitqtl.shape}

\item{shape}{A matrix of shape variables [n.ind n.pheno]}
}
\value{
The function returns the effect size of qtls.
}
\description{
Compute the effect size of qtl effect.
}
\details{
Function takes a qtl object, and an optionally a shape matrix.
}
\note{
If a shape matrix is done, then the function returns a list with the percentage explained of variance of the projection scores.
}
\examples{
qtl.effect <- fitqtlShape(cross, pheno.col=1:2, qtl, covar=NULL, "y~covar")
ES <- effectsizeShape()
}
\author{
Nicolas Navarro
}

