% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qtleffect_shape.R
\name{fitqtlShape}
\alias{fitqtlShape}
\title{fitqtlShape}
\usage{
fitqtlShape(cross, pheno.col, qtl, covar = NULL, formula, method = "hk",
  model = "mvnorm", dropone = FALSE, get.ests = TRUE, run.checks = TRUE,
  forceXcovar = FALSE, pca = NULL)
}
\arguments{
\item{cross}{A cross object.}
}
\value{
The function returns the qtl shape effects, eventually projected back in the tangent space.
}
\description{
fitqtlShape is a function that allows estimating qtl effect.
}
\note{
If cross is ....
}
\examples{
eff <- fitqtlShape(cross, pheno.col=1:2, qtl)
}
\author{
Nicolas Navarro
}
\references{
xxxx
}
\keyword{effect}
\keyword{shape}

