% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils_shape.R
\name{getGeno}
\alias{getGeno}
\title{getGeno}
\usage{
getGeno(cross, Q, add.only = FALSE)
}
\arguments{
\item{cross}{A cross object containing genotypes and some phenotypes}

\item{Q}{A data.frame containing chr and pos (similar to the output of summary.scanone)}

\item{add.only}{An optional argument (Default: FALSE). If f2 cross, add.only=TRUE forces the output to be the expected number of B alleles (pure additive model)}
}
\value{
Function returns the genotype with updated phenotypes.
}
\description{
Get genotypes in a cross at qtl positions
}
\details{
Function takes a cross object and a data.frame with 'chr' and 'pos' and returns genotype probabilities at the closest position in the genetic map.
}
\examples{
data(fake.bc)
fake.bc <- calc.genoprob(fake.bc)
Q <- data.frame(chr=1, pos=26)
geno <- getGeno(fake.bc, Q)
}
\author{
Nicolas Navarro
}
\keyword{utilities}

