% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.shapeEffect}
\alias{plot.shapeEffect}
\title{2D Lollipop plot of qtl effect}
\usage{
\method{plot}{shapeEffect}(mshape, effect, scaling = 1, links, labels,
  col.links, ...)
}
\arguments{
\item{mshape}{The mean shape. An array of [n.landmarks n.dim]}

\item{effect}{An array [n.landmarks n.dim] containing the effect (shape changes)
defined as the difference between a target shape and the mean shape}

\item{scaling}{Scaling factor. The target shape = mean shape + effect * scaling}

\item{links}{An optional matrix defining links between landmarks}
}
\description{
Plots scaled shape changes
}
\details{
Function takes the mean shape, a shape change, and a scaling factor and plots the shape changes as lollipop graph. The function plots either two- or three-dimensional shapes.
}
\examples{
x <- matrix(rnorm(10*3*2), nrow=10)
x <- asShapeArray(x, n.land = 3, n.dim = 2)
mshape <- matrix(c(0,0,1,0,0.5,1), nrow = 3, byrow = TRUE)
plot.shapeEffect(mshape, x[, , 1], scaling = 2, main="Triangle 1")
}
\author{
Nicolas Navarro
}

