% Generated by roxygen2 (4.0.1): do not edit by hand
\name{scanone.shape}
\alias{scanone.shape}
\title{Run one-dimensional genomescan}
\usage{
scanone.shape(cross, chr, pheno.col, model = "mvnorm", method = "hk",
  addcovar = NULL, intcovar = NULL, use = "all.obs", n.perm,
  perm.Xsp = FALSE, perm.strata = NULL, verbose = FALSE,
  batchsize = 250, n.cluster = 1, test = c("Pillai", "Hotelling.Lawley",
  "Lik.ratio"), formula)
}
\arguments{
\item{cross}{A cross object containing genotypes and some phenotypes}

\item{chr}{(Optional) Names of chromosomes use for mapping}

\item{pheno.col}{A numerical vector with column indicies of the phenotypes to map in cross$pheno or a vector containing their names. If missing, All but ID column.}

\item{model}{The phenotype model. Default and only possible value: \"mvnorm\"}

\item{method}{Method to use. Default and only value: \"hk\" for Haley-Knott regression}

\item{addcovar}{A data.frame with the additive covariates}

\item{intcovar}{A data.frame with the interactive covariates - KEEP IT DEFAULT, not use so far}

\item{use}{Only \"all.obs\" - KEEP IT DEFAULT, not use so far}

\item{n.perm}{Number of permutations}

\item{perm.Xsp}{- KEEP IT DEFAULT, not use so far}

\item{perm.strata}{- KEEP IT DEFAULT, not use so far}

\item{verbose}{- KEEP IT DEFAULT, not use so far}

\item{n.cluster}{Number of clusters for parallelization of permutations}

\item{batchsize}{- KEEP IT DEFAULT, not use so far}

\item{test}{Multivariate test statistics: \"Pillai\", \"Hotelling.Lawley\",\"Lik.ratio\". Allows partial matching.}

\item{formula}{Provides formula for the null model. Optional.}
}
\value{
Function returns one dimensional scan simimilar to the R/qtl \code{\link{scanone}} function.
}
\description{
A function to run QTL mapping on multivariate phenotypes (eg tangent coordinates or PCs)
}
\details{
Function takes a modified cross object with an additional class shape obtained by example with \code{\link{update.cross}}.
Additional covariates may be provided as supplementary arguments. Their ordering must match to the new phenotypes.
}
\examples{
data(fake.bc)
cross <- update.cross(cross, tgCoords, phen2keep, phen2update)
}
\author{
Nicolas Navarro
}
\seealso{
\code{\link{scanone}}
}
\keyword{utilities}

