% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scanone_shape.R
\name{scanoneShape}
\alias{scanoneShape}
\title{scanoneShape}
\usage{
scanoneShape(cross, chr, pheno.col, addcovar = NULL, intcovar = NULL,
  n.perm, perm.Xsp = FALSE, perm.strata = NULL, n.cluster = 1,
  test = c("Pillai", "Hotelling.Lawley", "Lik.ratio", "GoodallF"), formula)
}
\arguments{
\item{cross}{A cross object containing genotypes and some phenotypes}

\item{chr}{(Optional) Names of chromosomes use for mapping}

\item{pheno.col}{A numerical vector with column indicies in cross$pheno of
the phenotypes to map or a vector containing their names.
If missing, All but \"ID\" columns will be chosen as phenotypes.}

\item{addcovar}{A dataframe with the additive covariates}

\item{intcovar}{A dataframe with the interactive covariates - KEEP IT DEFAULT
This argument and the followings except if state otherwise are not use so far
They are there for direct}

\item{n.perm}{Number of permutations}

\item{perm.Xsp}{- KEEP IT DEFAULT, not use so far}

\item{perm.strata}{- KEEP IT DEFAULT, not use so far}

\item{n.cluster}{Number of clusters for parallelization of permutations}

\item{test}{Multivariate test statistics: \'Pillai\', 'Hotelling.Lawley',
\"Lik.ratio\", \"GoodallF\". Allows partial matching.}

\item{formula}{Provides formula for the null model. Optional.}
}
\value{
Function returns a data.frame similar to the R/qtl \code{\link{scanone}} function
with chromosome and cM positions in the two firsts columns and the $log_{10}$ of the p-value
for the multivariate phenotype. In the case of f2, there are three columns: full, purely additive and dominance models.
}
\description{
Runs genome scan on multivariate phenotypes with possible covariates
}
\details{
Function takes a modified cross object with an additional class shape
obtained from \code{\link{update.cross}}.
Additional covariates may be provided as supplementary arguments. Ordering of
observations in those additional covariates must match the new phenotypes.
}
\examples{
data(fake.bc)
fake.bc <- calc.genoprob(fake.bc, step=2.5)
out <- scanoneShape(fake.bc, chr = 1, pheno.col = 1:2, addcovar = getsex(cross)$sex,
test = "Pillai")
plot(out)
-----------
Permutations
perm.scan <- scanoneShape(fake.bc, pheno.col=1:2, addcovar = getsex(cross)$sex, n.perm=1000, n.cluster=2)
}
\author{
Nicolas Navarro
}
\seealso{
\code{\link{scanone}}
}
\keyword{analysis}

