% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stepwiseqtl_shape.R
\name{stepwiseqtlShape}
\alias{stepwiseqtlShape}
\title{stepwiseqtlShape}
\usage{
stepwiseqtlShape(cross, chr, pheno.col = 1, qtl, formula, max.qtl = 10,
  covar = NULL, refine.locations = TRUE, additive.only = FALSE,
  scan.pairs = FALSE, penalties, keeplodprofile = TRUE, keeptrace = FALSE,
  verbose = FALSE, test = "Pillai")
}
\arguments{
\item{cross}{}

\item{chr}{}

\item{pheno.col}{}

\item{qtl}{\code{\link[qtl]{summary.scanone}} object or any data.frame with 'chr' 'pos' and
'lod' values. If missing a one-dimensional scan is run first.}

\item{formula}{}

\item{max.qtl}{}

\item{covar}{}

\item{refine.locations}{Logical}

\item{additive.only}{Logical. (Defauft FALSE).
(Not as in \code{\link[qtl]{summary.scanone}}). Here setting to TRUE may be used
with f2 cross to force a diplotype model (pure additive)}

\item{scan.pairs}{Logical. (Default FALSE) #keep default. Not use so far}

\item{penalties}{Genomewide threshold}

\item{keeplodprofile}{Logical (Default TRUE)}

\item{keeptrace}{Logical (Default FALSE)}

\item{verbose}{Logical (Default FALSE)}

\item{test}{Multivariate statistics (Default "Pillai". Others are : "Hotelling.Lawley",
"Lik.ratio".)}
}
\value{
The function returns .
}
\description{
Runs a stepwise model search
}
\details{
The function built multiple QTL models using penalized negative $log_{10}$ of the p-value (as a multivariate analogue of the penalized LOD score of the R/qtl package). The search is limited so far to a model without interaction
}
\examples{
?stepwiseqtlShape
}
\author{
Nicolas Navarro
}
\references{
Broman and Sen 2009. A guide to QTL mapping with R/qtl
}
\seealso{
\code{\link[qtl]{stepwiseqtl}}
}

