% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils_shape.R
\name{update.cross}
\alias{update.cross}
\title{update.cross}
\usage{
\method{update}{cross}(cross, new.pheno = NULL, phen2keep = NULL,
  phen2update = NULL, id.geno = NULL, na.rm = FALSE, ...)
}
\arguments{
\item{cross}{A cross object containing genotypes and some phenotypes}

\item{new.pheno}{An optional data.frame containing new phenotypes}

\item{phen2keep}{An optional vector with names of the original phenotypes to keep}

\item{phen2update}{An optional vector with names of the subset of new
phenotypes to keep}

\item{id.geno}{An optional vector with the matching id in the genotypes}

\item{na.rm}{An optional logical if individuals with missing phenotypes
should be removed from the cross}
}
\value{
Function returns the cross object with updated phenotypes.
}
\description{
Update phenotypes and covariates in a cross qtl object. New phenotypes (for ex. tangent coordinates) are merged to an existing cross object obtained from \code{\link{read.cross}}
}
\details{
Function takes a cross object and a dataframe with new phenotypes (tangent coordinates, PC scores,...), matches cross$pheno$ID to the id column of the new phenotypes. Matching id.geno may be provide as an
optional input parameter.
Additional covariates may be provided as supplementary arguments but their ordering must match to the new phenotypes.
}
\examples{
data(fake.bc)
tgCoords <- data.frame(1:nind(fake.bc), rnorm(nind(fake.bc)), rnorm(nind(fake.bc)))
colnames(tgCoords) <- c('Id', 'ProcCoord1','ProcCoord2')
fake.bc <- update.cross(fake.bc, new.pheno=tgCoords, phen2keep=c("sex","age"), phen2update=colnames(tgCoords)[grep("ProcCoord",colnames(tgCoords))])
}
\author{
Nicolas Navarro
}
\seealso{
\code{\link{read.cross}}
}
\keyword{utilities}

