% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shapeQTL_support.R
\docType{data}
\name{wings.f2}
\alias{wings.f2}
\title{Landmark data from a simulated F2 of drosophila wings}
\description{
Landmark data from a simulated F2 of drosophila wings
}
\author{
Nicolas Navarro
}
\keyword{datasets}

