% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shapeQTL_support.R
\docType{data}
\name{wings}
\alias{wings}
\title{Landmark data from simulated drosophila wings}
\description{
Landmark data from simulated drosophila wings
}
\author{
Nicolas Navarro
}
\keyword{datasets}

