% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qtleffect_shape.R
\name{asShapeArray}
\alias{asShapeArray}
\title{asShapeArray}
\usage{
asShapeArray(shapes, n.land, n.dim, byrow = TRUE)
}
\arguments{
\item{shapes}{A matrix of shape coordinates n.obs rows of [x1 y1 {z1} x2 y2 {z2} ...]}

\item{n.land}{Number of landmarks}

\item{n.dim}{Number of dimensions (2 or 3)}

\item{byrow}{logical (default TRUE: data arranged initially according to x1 y1 z1 x2 y2 z2...). FALSE: data arranged initially according to x1 x2 x3 .... xp y1 .... yp z1 .... zp (used in the Morpho package)}
}
\value{
The function returns a multidimensional array [n.land x n.dim x n.obs]. The third dimension contains the names of each specimens if it was specified as row names in the input matrix.
}
\description{
Convert a matrix of landmark coordinates into an array [n.land x n.dim x n.obs].
}
\details{
The function converts a matrix of landmark coordinates into an array. Such arrays are common format used in other morphometrics packages (shape, geomorph, Morpho). Each row of the input matrix contains coordinates for a single specimens.
}
\examples{
x <- matrix(rnorm(5*3*2),nrow=5, byrow=TRUE) # 5 random triangles
asShapeArray(x, 3, 2, byrow = TRUE)
}
\author{
Nicolas Navarro
}
\seealso{
\code{\link[geomorph]{arrayspecs}}
}
\keyword{utilities}

