% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qtleffect_shape.R
\name{effectsizeShape}
\alias{effectsizeShape}
\title{Effect size for multivariate shape vector}
\usage{
effectsizeShape(qtl, shape = NULL, geno = NULL, ...)
}
\arguments{
\item{qtl}{A qtl effect obtained from fitqtlShape}

\item{shape}{A optional matrix of shape variables [n.ind n.pheno]}

\item{geno}{A optional matrix of genotype data [n.ind n.qtl]}
}
\value{
The function returns the effect size of qtls.
}
\description{
Compute the effect size of qtl effect.
}
\details{
Function takes a qtl object, and an optionally a shape matrix.
}
\note{
If the shape matrix and/or the genotypes are provided, then the function returns a dataframe with the ES standardized by genotype variance and the percentage explained of variance of the projection scores.
}
\examples{
ES <- effectsizeShape(qtl, shape, geno = as.matrix(getGeno(cross, qtl)))
}
\author{
Nicolas Navarro
}

