% Generated by roxygen2 (4.0.2): do not edit by hand
\name{effectsizeShape}
\alias{effectsizeShape}
\title{Effect size for multivariate shape vector}
\usage{
effectsizeShape(qtl, shape = NULL)
}
\arguments{
\item{qtl}{A qtl effect obtained from fitqtl.shape}

\item{shape}{A matrix of shape variables [n.ind n.pheno]}
}
\value{
The function returns the effect size of qtls.
}
\description{
A function to compute the effect size from the result obtained from fitqtl.shape.
}
\details{
Function takes the qtl object, an optionally a shape matrix.
}
\note{
If a shape matrix is done, then the function returns a list with the percentage explained of variance of the projection scores.
}
\examples{
qtl.effect <- fitqtl.shape(cross, pheno.col=1:2, qtl, covar=NULL, "y~covar")
ES <- effect.size()
plot.effect(m.shape,target,n.dim,n.land,main="Q1")
}
\author{
Nicolas Navarro
}
\references{
xxxx
}
\seealso{
\code{\link{fitqtlShape}}
}
\keyword{effect}
\keyword{size}

