% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fitqtlShape}
\alias{fitqtlShape}
\title{fitqtlShape}
\usage{
fitqtlShape(cross, pheno.col, qtl, covar = NULL, formula, method = "hk",
  model = "mvnorm", dropone = FALSE, get.ests = TRUE, run.checks = TRUE,
  forceXcovar = FALSE, pca = NULL)
}
\arguments{
\item{cross}{A cross object.}
}
\value{
The function returns the qtl shape effects, eventually projected back in the tangent space.
}
\description{
fitqtl.shape is a function that allows estimating qtl effect.
}
\note{
If cross is ....
}
\examples{
eff <- fitqtl.shape(cross, pheno.col=1:2, qtl, covar=NULL, "y~covar")
}
\author{
Nicolas Navarro
}
\references{
xxxx
}
\seealso{
\code{\link{fitqtl}}
}
\keyword{effect}
\keyword{shape}

