% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils_shape.R
\name{getGeno}
\alias{getGeno}
\title{Get genotypes in a cross at qtl positions}
\usage{
getGeno(cross, Q, add.only = FALSE)
}
\arguments{
\item{cross}{A cross object containing genotypes and some phenotypes}

\item{Q}{A data.frame containing chr and pos (similar to the output of summary.scanone)}

\item{add.only}{An optional argument (Default: FALSE). TRUE force the output to the probability of the number of B alleles in F2}
}
\value{
Function returns the genotype with updated phenotypes.
}
\description{
A function to get genotypes in a cross
}
\details{
Function takes a cross object and summary.scanone object, matches cross$pheno$id to
the id column of the new phenotypes. Matching id.geno may be provide as an optional input parameter.
Additional covariates may be provided as supplementary arguments. Their ordering must match to the new phenotypes.
}
\examples{
data(fake.bc)
fake.bc <- calc.genoprob(fake.bc)
Q <- data.frame(chr=1,pos=26)
geno <- getGeno(fake.bc, Q)
}
\author{
Nicolas Navarro
}
\keyword{utilities}

