% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.shapeEffect}
\alias{plot.shapeEffect}
\title{2D Lollipop plot of qtl effect}
\usage{
\method{plot}{shapeEffect}(mshape, target, n.dim, n.land, links, main = NULL,
  ...)
}
\arguments{
\item{mshape}{The mean shape. An array of [n.landmarks n.dim]}

\item{target}{An array [n.landmarks n.dim] containing the vector of shape changes}

\item{links}{An optional matrix defining links between landmarks}
}
\description{
A function to plot scaled qtl effect from the result obtained from fitqtl.shape.
}
\details{
Function takes the mean shape, a target shape, and a scaling factor and plot the shape changes as lollipop graph. The function plot either two- or three-dimensional shapes.
}
\examples{
qtl.effect <- fitqtl.shape(cross, pheno.col=1:2, qtl, covar=NULL, "y~covar")
target <- qtl.effect*scaling + m.shape
plot.shape(m.shape,target,n.dim,n.land,main="Q1")
}
\author{
Nicolas Navarro
}
\seealso{
\code{\link{fitqtl.shape}}
}
\keyword{plot}

