% Generated by roxygen2 (4.0.2): do not edit by hand
\name{scanoneShape}
\alias{scanoneShape}
\title{Run one-dimensional genomescan}
\usage{
scanoneShape(cross, chr, pheno.col, addcovar = NULL, intcovar = NULL,
  n.perm, perm.Xsp = FALSE, perm.strata = NULL, n.cluster = 1,
  test = c("Pillai", "Hotelling.Lawley", "Lik.ratio"), formula)
}
\arguments{
\item{cross}{A cross object containing genotypes and some phenotypes}

\item{chr}{(Optional) Names of chromosomes use for mapping}

\item{pheno.col}{A numerical vector with column indicies in cross$pheno of
the phenotypes to map or a vector containing their names.
If missing, All but \"ID\" columns will be chosen as phenotypes.}

\item{addcovar}{A dataframe with the additive covariates}

\item{intcovar}{A dataframe with the interactive covariates - KEEP IT DEFAULT
This argument and the followings except if state otherwise are not use so far
They are there for direct}

\item{n.perm}{Number of permutations}

\item{perm.Xsp}{- KEEP IT DEFAULT, not use so far}

\item{perm.strata}{- KEEP IT DEFAULT, not use so far}

\item{n.cluster}{Number of clusters for parallelization of permutations}

\item{test}{Multivariate test statistics: \"Pillai\", \"Hotelling.Lawley\",
\"Lik.ratio\". Allows partial matching.}

\item{formula}{Provides formula for the null model. Optional.}
}
\value{
Function returns one dimensional scan similar to the R/qtl \code{\link{scanone}} function.
}
\description{
A function to run QTL mapping on multivariate phenotypes
}
\details{
Function takes a modified cross object with an additional class shape
obtained from \code{\link{update.cross}}.
Additional covariates may be provided as supplementary arguments. Ordering of
observations in those additional covariates must match the new phenotypes.
}
\examples{
data(fake.bc)
out <- scanoneShape(cross, chr = 1, pheno.col = 1:2, addcovar = getsex(cross)$sex,
test = "Pillai")
}
\author{
Nicolas Navarro
}
\seealso{
\code{\link{scanone}}
}
\keyword{qtl,}
\keyword{shape}

